# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Japan Eco Tracking API'
  version: 1.0.0
paths:
...
  '/api/v1/area-photos/{id}':
    delete:
      tags:
        - AreaPhotos
      summary: 'Delete a area'
      description: 'Delete a area by id'
      operationId: Delete
      parameters:
        -
          name: id
          in: path
          description: 'ID of areaPhoto to return'
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        200:
          description: 'Delete areaPhoto success'
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                    example: '200'
                  message:
                    type: string
                    example: 'Delete areaPhoto success'
                  data:
                    properties:
                      id:
                        type: integer
                        example: '1'
                      name:
                        type: email
                        example: example
                      created_at:
                        type: string
                        example: null
                      updated_at:
                        type: string
                        example: null
                    type: object
                type: object
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: 'Member sign in'
      description: 'Login by email, password'
      operationId: login
      requestBody:
        description: 'Pass user credentials'
        required: true
        content:
          application/json:
            schema:
              required:
                - username
                - password
                - type
                - device_token
              properties:
                username:
                  type: string
                  example: username
                password:
                  type: string
                  example: password
                type:
                  type: string
                  example: member
                device_token:
                  type: string
                  example: AHWR3LRhNIiotrpfJtK1
                device_agent:
                  type: string
                  example: 'Mozilla/5.0 (Windows 95) AppleWebKit/5322 (KHTML, like Gecko) Chrome/39.0.866.0 Mobile Safari/5322'
              type: object
      responses:
        200:
          description: 'Login success'
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                    example: '200'
                  message:
                    type: string
                    example: 'Login success'
                  data:
                    properties:
                      member_id:
                        type: string
                        example: 1
                      access_token:
                        type: string
                        example: 7b4267b17e12512af1ddc3a2cc313e3a7f2b3d22dbcb1b92cb127cf6596a15443a140863452ad4e9ae850f288921c5dff61bba6f3f422fb850942023956eb969
                      token_type:
                        type: string
                        example: Bearer
                      device_token:
                        type: string
                        format: type
                        example: AHWR3LRhNIiotrpfJtK1
                      device_agent:
                        type: string
                        format: type
                        example: 'Mozilla/5.0 (Windows 95) AppleWebKit/5322 (KHTML, like Gecko) Chrome/39.0.866.0 Mobile Safari/5322'
                    type: object
                type: object
  /api/v1/auth/logout:
    get:
      tags:
        - Authentication
      summary: Logout
      description: Logout
      operationId: logout
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            type: string
          example: 76f58b2747fa65b4dc157f9c1c9bbff708bffac652d8f67a60cff95ef4faf956e2767846b811af4133e277a1664f195853864b911c87c3c6fcf4e6bcccaedb86
      responses:
        200:
          description: 'Logout success'
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                    example: '200'
                  message:
                    type: string
                    example: 'Logout success'
                  data:
                    type: string
                    example: true
                type: object
...
components:
...
  securitySchemes:
    bearerAuth:
      type: apiKey
      description: 'A token created from Sanctum Laravel package'
      name: Authorization
      in: header
security:
  - []